Refactoring to make the SystemTypeSelection component shareable
