+++ src/page-multi-facility/LocaleInformationSection.tsx
@@ -4,13 +4,9 @@ import styled from "styled-components";
 import InputSelect from "../design-system/InputSelect";
 import LocaleInformation from "../impact-dashboard/LocaleInformation";
 import { SectionHeader } from "./FacilityInputForm";
+import SystemTypeSelection from "./SystemTypeSelection";
 
 const LocaleInformationSectionDiv = styled.div``;
-
-const SystemTypeInputDiv = styled.div`
-  width: 195px;
-  padding-bottom: 20px;
-`;
 
 interface Props {
   systemType?: string;
@@ -21,28 +17,13 @@ const LocaleInformationSection: React.FC<Props> = ({
   systemType,
   setSystemType,
 }) => {
-  const systemTypeList = [
-    { value: undefined },
-    { value: "State Prison" },
-    { value: "County Jail" },
-  ];
-
   return (
     <LocaleInformationSectionDiv>
       <SectionHeader>Locale Information</SectionHeader>
-      <SystemTypeInputDiv>
-        <InputSelect
-          label="Type of System"
-          value={systemType}
-          onChange={(event) => setSystemType(event.target.value)}
-        >
-          {systemTypeList.map(({ value }, index) => (
-            <option key={value || `system-type-${index}`} value={value}>
-              {value}
-            </option>
-          ))}
-        </InputSelect>
-      </SystemTypeInputDiv>
+      <SystemTypeSelection
+        systemType={systemType}
+        setSystemType={setSystemType}
+      />
       <LocaleInformation />
     </LocaleInformationSectionDiv>
   );
